iml/exercises/w01/ex01/requirements.txt
pytorch::pytorch
pytorch::torchvision
conda-forge::torchinfo
numpy
jupyter
matplotlib
Pillow
scikit-learn